Code Restructure and Testing
Restructured the code and added pytest testing.
