+++ src/tests/test_assignment2.py
@@ -1,0 +1,8 @@
+import sys
+ 
+# append the path of the parent directory
+sys.path.append("..")
+
+from BMI import BMICalc
+
+BMICalc.test_bmicalc()
